port computeForces on device
