+++ arcane/samples_build/samples/hydro_accelerator/SimpleHydro.axl
@@ -463,6 +463,9 @@
       </description>
     </simple>
     
+    <simple name = "use-multiple-queue-for-boundary-conditions" type="bool" default="false" >
+       <description>Indique si on utilise plusieurs une file d'exécution par condition aux limites</description>
+    </simple>
     
     <!-- - - - - - func order - - - - -->
     <simple name="func-order" type="integer" default="0" >
